requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.1
torchvision==0.16.1
numpy==1.26.2
pillow==10.1.0
opencv-python==4.8.1.78
pydicom==2.4.3
python-multipart==0.0.6
transformers==4.35.2
scikit-learn==1.3.2
pandas==2.1.3 